fix api ログインapiの定義を統合
