+++ apps/api/src/user/dto/get-user.dto.ts
@@ -1,7 +1,13 @@
 import { ApiProperty } from "@nestjs/swagger";
 import { MaxLength, Matches } from "class-validator";
 
-export class GetUserDto {
+export type User = {
+  id: string;
+  email: string;
+  name: string;
+};
+
+export class GetUserRequest {
   @ApiProperty({
     example: "ckv9ydh6s0000gkpj1wybug0x",
     description: "ユーザーID",
@@ -9,4 +15,16 @@ export class GetUserDto {
   @MaxLength(36)
   @Matches(/^c[a-z0-9]+$/)
   id!: string;
+}
+
+export class GetUserResponse {
+  @ApiProperty({
+    example: {
+      id: "ckv9ydh6s0000gkpj1wybug0x",
+      email: "user@example.com",
+      name: "John Doe",
+    },
+    description: "ユーザー情報",
+  })
+  user!: User;
 }
